# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Generate Admin",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/session": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "Read Session",
        "operationId": "readSession",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/session/accept-invitation": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Accept Invitation",
        "operationId": "acceptInvitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/session/access-request": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Request Access",
        "operationId": "requestAccess",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRequestRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInvitation": {
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token"
        ],
        "title": "AcceptInvitation"
      },
...
      "AccessRequestCreate": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AccessRequestCreate"
      },
...
      "AccessRequestRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "entra_object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entra Object Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "status": {
            "$ref": "#/components/schemas/AccessRequestStatus"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "entra_object_id",
          "name",
          "message",
          "status",
          "reviewed_by",
          "reviewed_at",
          "decision_note",
          "created_at"
        ],
        "title": "AccessRequestRead"
      },
      "AccessRequestStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "denied"
        ],
        "title": "AccessRequestStatus"
      },
      "AccessState": {
        "type": "string",
        "enum": [
          "no_access",
          "invited",
          "pending",
          "denied",
          "active",
          "no_roles",
          "suspended"
        ],
        "title": "AccessState"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Identity": {
        "properties": {
          "entra_object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entra Object Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "entra_object_id",
          "email",
          "name"
        ],
        "title": "Identity"
      },
...
      "RoleAssignmentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/RoleSummary"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "granted_at",
          "expires_at"
        ],
        "title": "RoleAssignmentRead"
      },
...
      "RoleSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "name"
        ],
        "title": "RoleSummary"
      },
      "Session": {
        "properties": {
          "access_state": {
            "$ref": "#/components/schemas/AccessState"
          },
          "identity": {
            "$ref": "#/components/schemas/Identity"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "access_state",
          "identity"
        ],
        "title": "Session"
      },
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "entra_object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entra Object Id"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "role_assignments": {
            "items": {
              "$ref": "#/components/schemas/RoleAssignmentRead"
            },
            "type": "array",
            "title": "Role Assignments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "status",
          "entra_object_id",
          "last_login_at",
          "created_at"
        ],
        "title": "UserRead"
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended"
        ],
        "title": "UserStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}