# packages/api/src/generated/model/accessRequestStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Generate Admin
 * OpenAPI spec version: 0.1.0
 */

export type AccessRequestStatus = typeof AccessRequestStatus[keyof typeof AccessRequestStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AccessRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  denied: 'denied',
} as const;