# packages/api/src/generated/model/accessState.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Generate Admin
 * OpenAPI spec version: 0.1.0
 */

export type AccessState = typeof AccessState[keyof typeof AccessState];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AccessState = {
  no_access: 'no_access',
  invited: 'invited',
  pending: 'pending',
  denied: 'denied',
  active: 'active',
  no_roles: 'no_roles',
  suspended: 'suspended',
} as const;