# packages/api/src/generated/model/identity.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Generate Admin
 * OpenAPI spec version: 0.1.0
 */

export interface Identity {
  entra_object_id: string;
  email: string;
  name: string;
}