# packages/api/src/generated/model/roleAssignmentRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Generate Admin
 * OpenAPI spec version: 0.1.0
 */
import type { RoleSummary } from './roleSummary';
import type { RoleAssignmentReadExpiresAt } from './roleAssignmentReadExpiresAt';

export interface RoleAssignmentRead {
  id: string;
  role: RoleSummary;
  granted_at: string;
  expires_at: RoleAssignmentReadExpiresAt;
}