# packages/api/src/generated/model/userRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Generate Admin
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatus } from './userStatus';
import type { UserReadLastLoginAt } from './userReadLastLoginAt';
import type { RoleAssignmentRead } from './roleAssignmentRead';

export interface UserRead {
  id: string;
  email: string;
  name: string;
  status: UserStatus;
  entra_object_id: string;
  last_login_at: UserReadLastLoginAt;
  created_at: string;
  role_assignments?: RoleAssignmentRead[];
}